requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
djoser
pymysql
djangorestframework_simplejwt
django-colorfield